+++ xlators/mgmt/glusterd/src/glusterd-sm.h
@@ -104,7 +104,7 @@ struct glusterd_peerinfo_ {
         /* Members required for proper cleanup using RCU */
         gd_rcu_head                     rcu_head;
         pthread_mutex_t                 delete_lock;
-        uint64_t                        generation;
+        uint32_t                        generation;
 };
 
 typedef struct glusterd_peerinfo_ glusterd_peerinfo_t;
